requirements.txt
Flask
google-cloud-storage
werkzeug